refactor(vscode): inline setupPreviewBehavior into <preview-app>.firstUpdated
Verbatim move of `setupPreviewBehavior`'s body from `behavior.ts` into
`<preview-app>.firstUpdated`. Closure shape, controller construction
order, event listeners, and message dispatch are preserved. The
`behavior.ts` module is deleted as the function is no longer a separate
export. No behaviour change.

Step 1 of #860; subsequent steps will introduce @query handles, lift
local `let` bindings to component state, migrate controllers to
ReactiveControllers, and lift message dispatch.
